Move driver definition to extension + allow executing methods
