+++ src/vs/workbench/services/positronConnections/browser/positronConnectionsInstance.ts
@@ -6,13 +6,13 @@
 import { Disposable } from '../../../../base/common/lifecycle.js';
 import { Emitter, Event } from '../../../../base/common/event.js';
 import { ConnectionsClientInstance } from '../../languageRuntime/common/languageRuntimeConnectionsClient.js';
-import { ConnectionMetadata, IPositronConnectionInstance, IPositronConnectionItem } from './interfaces/positronConnectionsInstance.js';
+import { ConnectionMetadata, IPositronConnectionInstance, IPositronConnectionItem, IPositronConnectionEntry } from '../common/interfaces/positronConnectionsInstance.js';
 import { ObjectSchema } from '../../languageRuntime/common/positronConnectionsComm.js';
 import { IRuntimeSessionService } from '../../runtimeSession/common/runtimeSessionService.js';
 import { RuntimeCodeExecutionMode, RuntimeErrorBehavior } from '../../languageRuntime/common/languageRuntimeService.js';
-import { flatten_children, IPositronConnectionEntry } from './positronConnectionsUtils.js';
+import { flatten_children } from './positronConnectionsUtils.js';
 import { Severity } from '../../../../platform/notification/common/notification.js';
-import { IPositronConnectionsService } from './interfaces/positronConnectionsService.js';
+import { IPositronConnectionsService } from '../common/interfaces/positronConnectionsService.js';
 import { DeferredPromise } from '../../../../base/common/async.js';
 import { localize } from '../../../../nls.js';
