+++ src/vs/workbench/services/positronConnections/common/interfaces/positronConnectionsService.ts
@@ -9,7 +9,7 @@ import { Emitter, Event } from '../../../../../base/common/event.js';
 import Severity from '../../../../../base/common/severity.js';
 import { INotificationHandle } from '../../../../../platform/notification/common/notification.js';
 import { IRuntimeSessionService } from '../../../runtimeSession/common/runtimeSessionService.js';
-import { PositronConnectionsDriverManager } from '../positronConnectionsDrivers.js';
+import { PositronConnectionsDriverManager } from '../../browser/positronConnectionsDrivers.js';
 
 export const IPositronConnectionsService = createDecorator<IPositronConnectionsService>('positronConnectionsService');
 export const POSITRON_CONNECTIONS_VIEW_ID = 'workbench.panel.positronConnections';
